feat: add centralized state selectors
